chatbot-api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
groq==0.25.0
requests==2.32.3
beautifulsoup4==4.13.4
python-dotenv==1.1.0
